Related to #178 

**Changes:**
- purr/compiler/devices.py: caching `id` and `full_id`, deleting the cached ids
- purr/compiler/hardware_models.py: `delete_cache()` to clean the cached ids for all components,
- Tests to ensure caching works as intended...

**A few things to consider (please give your thoughts):**
- The cached `full_id` is still retrieved using a function call: this shouldn't break any dependencies, but obviously has some overhead. Would be happy to change.
- I'm of the opinion that the architecture of hardware models should be immutable (including ids) but have calibratables that can be changed and having the functionality to change IDs like I've added here shouldn't be encouraged... However, right now, these do serve as a useful way to test that the clearing the cache does work as intended. I have added them just to the tests.

Some data on the kind of speeds up we see (why caching the ids is important): N qubits each with 100 cnot layers and 100 U layers.
|  | N=10 before | N=10 after | N=2 before | N=2 after |
| --- | --- | --- | --- | --- |
| Building the circuit | 0.434s | 0.312s | 0.0476s | 0.0396s |
| Optimising the circuit | 0.104s | 0.0352s | 0.0149s | 0.00627s |
| Creating the timeline | 0.332s | 0.156s | 0.0589s | 0.0280s |