@hamidelmaazouz 2024-09-23
Here's an example:

```
class SomeClass:
    def __init__(self, number):
        self._number = number

    def __eq__(self, other):
        return self._number == other._numer

    def __hash__(self):
        return hash(self._number)

    @property
    @lru_cache
    def doubled(self):
        return self.number * 2

    @property
    def number(self):
        return self._number

    @number.setter
    def number(self, val):
        self._number = val


obj = SomeClass(1)
assert obj.doubled == 2 * 1
assert obj.doubled == 2 * obj.number

obj.number = 5 # Access through a setter : Okay
assert obj.doubled == 2 * 5
assert obj.doubled == 2 * obj.number

obj._number = 2 # Access through a private field : Okay
assert obj.doubled == 2 * obj.number
assert obj.doubled == 2 * 2

```