@lcauser-oqc 2024-09-24
> Love that you've attempted to manage the cache yourself, sounds great but it overfits the complexity the whole problem deserves. I also don't want to introduce`cached_full_id` as a new member or inherit the headaches of managing either's cached or non cached property changes, all of that plays against simplicity and ease to debug if anything ever goes wrong.
> 
> Instead, have a look at:
> 
> * Last four sections of `@cached_property` https://docs.python.org/3/library/functools.html#functools.cached_property
> * `@lru_cache` and more useful info at https://docs.python.org/3/faq/programming.html#how-do-i-cache-method-calls
> * Seemingly unrelated, but checkout the story about thread safety of `@cached_property`at https://discuss.python.org/t/finding-a-path-forward-for-functools-cached-property/23757/17?page=2.

Thanks for the nice suggestions!

After some playing with it, I think there will be some problems with using `@lru_cache` that would require a solution as specific as the one we have here. The main difficulty is that caching the outcome of `full_id` will rely on on `__hash__`. In your example, `__hash__` only depends on `number` so it would make sense. Here, we need it to depend on `full_id`, which will get us stuck in a recursion loop. If we try to hash using something other than the `full_id`, e.g. the `partial_id`, then we risk not having full coverage to any possible changes to the `full_id` (in the case that we hash with `partial_id`, the cached `full_id` would not be updated if the `id` of the physical channel changes).