@hamidelmaazouz 2024-09-24
> Here, we need it to depend on full_id, which will get us stuck in a recursion loop. If we try to hash using something other than the full_id, e.g. the partial_id, then we risk not having full coverage to any possible changes to the full_id (in the case that we hash with partial_id, the cached full_id would not be updated if the id of the physical channel changes).

Good point on the recursion idea, have you tried it ?
Another simple solution is keep the @cached_property decorator, but enforce immutability all components composing the full_id