@hamidelmaazouz 2024-09-24
I am more in favour of enforcing immutability, but allow users to copy-construct pulse channels if they need to