@lcauser-oqc 2024-09-24
I did try using the `lru_cache`: changing the hash to the `partial_id` instead of the `full_id` did break the loop, but as anticipated, did not allow for the `full_id` to be updated if the id of a physical channel is changed. 

I am also in favour of enforcing immutability. However, there is need to dynamically change the hardware model (@bgsach might have details on this).